--- BulkDelete/pom.xml
@@ -1,15 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>com.oracle.kv.examples</groupId>
-  <artifactId>BulkDelete</artifactId>
-  <version>0.0.1-SNAPSHOT</version>
-  <name>Efficient Multi Delete</name>
-  <description>Bulk deletion of major keys when full major key list is unknown</description>
-  <dependencies>
-  	<dependency>
-  		<groupId>com.oracle.kv</groupId>
-  		<artifactId>oracle-nosql-client</artifactId>
-  		<version>3.0.14</version>
-  	</dependency>
-  </dependencies>
-</project>
