+++ examples-nosql-java-direct-driver/BulkDelete/README.md
