+++ examples-nosql-java-direct-driver/BulkGet/BulkGetExample.java
