+++ examples-nosql-java-direct-driver/BulkPut/BulkPutExample.java
