+++ examples-nosql-java-direct-driver/README.md
@@ -1,0 +1,1 @@
+# examples-nosql-java-direct-driver
