+++ examples-nosql-java-sdk/HelloWorld/HelloWorld.java
